% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1004_ire_noise.R
\name{ire.noise}
\alias{ire.noise}
\title{Title irregular_pattern with noise, ZINB or ZIP}
\usage{
ire.noise(
  noise = 0.1,
  lambda = 0.7,
  spots,
  se,
  ns,
  type,
  se.p,
  se.size = 0.2,
  se.mu,
  ns.p,
  ns.size = 0.2,
  ns.mu,
  ptn,
  png_dir
)
}
\arguments{
\item{noise}{swap porpotion.}

\item{lambda}{The probability of the sample() function process.}

\item{spots}{The number of spots.}

\item{se}{The number of  spatially variable genes (SVGs).}

\item{ns}{The number of  non-SVGs.}

\item{type}{A character.'ZINB' or 'ZIP. The default value is ZINB.}

\item{se.p}{A number. The zero proportion of zero generation process of
SVGs in the streak area.}

\item{se.size}{The size para of SVGs in the NB distribution.}

\item{se.mu}{For SVGs, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ns.p}{A number. The zero proportion of zero generation process of
non-SVGs and SVGs in the non-streak area.}

\item{ns.size}{For non-SVGs and SVGs in the non-streak area,
the size para in the NB distribution.}

\item{ns.mu}{For non-SVGs and SVGs in the non-streak area, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ptn}{The file name of the pattern png.}

\item{png_dir}{The dir of the png files.}
}
\value{
A data.frame of gene expression counts and coordinates.
}
\description{
irregular_pattern with noise, swap noise function 1, ZINB or ZIP, for irregular pattern.
}
