% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1017_mix_noise.R
\name{mix_ire}
\alias{mix_ire}
\title{Title For irregular pattern, Simulation data generation with mixture_noise.
Title Simulation data generation with mixture_noise.}
\usage{
mix_ire(lambda = lambda, mix = 0.1, spots, gns, type, ptn, png_dir)
}
\arguments{
\item{lambda}{The probability of the sample() function process.}

\item{mix}{the mixture proportion}

\item{spots}{The number of spots.}

\item{gns}{the number of SVGs，}

\item{type}{A character.'ZINB' or 'ZIP','Pois', or "NB".}

\item{ptn}{The file name of the pattern png.}

\item{png_dir}{The dir of the png files.}
}
\value{
a data frame
}
\description{
Title For irregular pattern, Simulation data generation with mixture_noise.
Title Simulation data generation with mixture_noise.
}
