% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1004new_noise_simu.R
\name{noise.zi2}
\alias{noise.zi2}
\title{Title swap noise function 3}
\usage{
noise.zi2(
  lambda = 0.7,
  spots,
  se,
  ns,
  type,
  se.p,
  se.size = 0.2,
  se.mu,
  ns.p,
  ns.size = 0.2,
  ns.mu,
  ptn,
  png_dir,
  noise = 0.1
)
}
\arguments{
\item{lambda}{The probability of the sample() function process.}

\item{spots}{The number of spots.}

\item{se}{The number of  spatially variable genes (SVGs).}

\item{ns}{The number of  non-SVGs.}

\item{type}{A character.'ZINB' or 'ZIP. The default value is ZINB.}

\item{se.p}{A number. The zero proportion of zero generation process of
SVGs in the streak area.}

\item{se.size}{The size para of SVGs in the NB distribution.}

\item{se.mu}{For SVGs, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ns.p}{A number. The zero proportion of zero generation process of
non-SVGs and SVGs in the non-streak area.}

\item{ns.size}{For non-SVGs and SVGs in the non-streak area,
the size para in the NB distribution.}

\item{ns.mu}{For non-SVGs and SVGs in the non-streak area, the lambda para in the poisson distribution
or the mu para in the NB distribution.}

\item{ptn}{The file name of the pattern png.}

\item{png_dir}{The dir of the png files.}

\item{noise}{swap porpotion.}
}
\value{
A data.frame of gene expression counts and coordinates.
}
\description{
The swap noise function3, swap noise, in the ZINB or ZIP.
The diff between SVG and non-SVG is large.
}
