% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1004_new_noisebasic.R
\name{simu_norm2}
\alias{simu_norm2}
\title{Title zero-inflated Normal distribution, without cut.}
\usage{
simu_norm2(subject.n, zi.p = 0.5, mu = 0.5, sd = 0.25)
}
\arguments{
\item{subject.n}{unused argument}

\item{zi.p}{unused argument}

\item{mu}{mu of Normal.}

\item{sd}{sd of Normal.}
}
\value{
a vector. The simulation data.
}
\description{
2023_1004 new simu function, generating simulation data from zero-inflated Normal distribution,
without the cut (threshold).
}
