% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1004_new_noisebasic.R
\name{simu_part_norm}
\alias{simu_part_norm}
\title{Title  对数正态的的norm-----}
\usage{
simu_part_norm(subject.n, zi.p = 0.5, mu = 0.5, sd = 0.25, cut = 5)
}
\arguments{
\item{subject.n}{unused argument}

\item{zi.p}{unused argument}

\item{mu}{mu of Normal.}

\item{sd}{sd of Normal.}

\item{cut}{the threshold.Ensuring that the expression of SVG has obviously high values.}
}
\value{
a vector. The simulation data.
}
\description{
2023_1004 new simu function, generating simulation data from normal Normal distribution.
2^(x) follow logNormal. Skewness，with cut(the threshold).
}
