% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1004_new_noisebasic.R
\name{simu_zi}
\alias{simu_zi}
\title{Title original simu_zi function}
\usage{
simu_zi(family, subject.n, zi.p = 0.5, mu = 0.5, size = 0.25)
}
\arguments{
\item{family}{ZINB or ZIP.}

\item{subject.n}{sample size.}

\item{zi.p}{zero-inflated proportion.}

\item{mu}{if ZINB, it is the mu parameter, if ZIP, it is the lambda paramter.}

\item{size}{if ZINB, it is the size parameter, if ZIP, it is the unused argument.}
}
\value{
a vector. The simulation data.
}
\description{
original simu_zi function, generating simualtion data from ZINB or ZIP distribution.
}
